requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
requests==2.32.3
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
flask-session==0.5.0
